% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossFit.R
\name{CrossFit}
\alias{CrossFit}
\title{Perform Cross-fitting}
\usage{
CrossFit(Tr, P, y, x, type = c("binomial", "continuous"), learners, folds)
}
\arguments{
\item{Tr}{training data}

\item{P}{prediction data}

\item{y}{outcome variable name}

\item{x}{covariate variable names}

\item{type}{outcome variable type (i.e, "binomial", "continuous")}
}
\description{
Perform Cross-fitting
}
