% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{.lcmmtp_control}
\alias{.lcmmtp_control}
\title{Set Estimation Hyperparameters}
\usage{
.lcmmtp_control(...)
}
\arguments{
\item{...}{}
}
\description{
Set Estimation Hyperparameters
}
\examples{
.lcmmtp_control(folds = 5, learners_trt = c("glm", "earth"))
}
