% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcm_Vars.R
\name{lcm_Vars}
\alias{lcm_Vars}
\title{R6 class for an lcm_Vars}
\description{
R6 class for an lcm_Vars

R6 class for an lcm_Vars
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-lcm_Vars-new}{\code{lcm_Vars$new()}}
\item \href{#method-lcm_Vars-history}{\code{lcm_Vars$history()}}
\item \href{#method-lcm_Vars-all_vars}{\code{lcm_Vars$all_vars()}}
\item \href{#method-lcm_Vars-clone}{\code{lcm_Vars$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lcm_Vars-new"></a>}}
\if{latex}{\out{\hypertarget{method-lcm_Vars-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lcm_Vars$new(W, L, A, Z, M, Y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lcm_Vars-history"></a>}}
\if{latex}{\out{\hypertarget{method-lcm_Vars-history}{}}}
\subsection{Method \code{history()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lcm_Vars$history(var = c("L", "A", "Z", "M", "Y"), t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lcm_Vars-all_vars"></a>}}
\if{latex}{\out{\hypertarget{method-lcm_Vars-all_vars}{}}}
\subsection{Method \code{all_vars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lcm_Vars$all_vars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lcm_Vars-clone"></a>}}
\if{latex}{\out{\hypertarget{method-lcm_Vars-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lcm_Vars$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
