% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmmtp-package.R
\docType{package}
\name{lcmmtp-package}
\alias{lcmmtp}
\alias{lcmmtp-package}
\title{lcmmtp: Nonparametric Longitudinal Causal Mediation Using Modified Treatment Policies}
\description{
Nonparametric longitudinal causal mediation using modified treatment policies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nt-williams/lcmmtp}
  \item Report bugs at \url{https://github.com/nt-williams/lcmmtp/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Williams \email{ntwilliams.personal@gmail.com} (\href{https://orcid.org/0000-0002-1378-4831}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Iván Díaz \email{ivan.diaz@nyulangone.org} (\href{https://orcid.org/0000-0001-9056-2047}{ORCID}) [copyright holder]
}

}
\keyword{internal}
