% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Npsem.R
\name{Npsem}
\alias{Npsem}
\title{Create an Npsem Object}
\description{
Create an Npsem Object

Create an Npsem Object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Npsem$new()}}
\item \href{#method-parents}{\code{Npsem$parents()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Npsem$new(W, L, A, Z, M, Y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parents"></a>}}
\if{latex}{\out{\hypertarget{method-parents}{}}}
\subsection{Method \code{parents()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Npsem$parents(var = c("L", "A", "Z", "M", "Y"), t)}\if{html}{\out{</div>}}
}

}
}
