% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmmtp.R
\name{lcmmtp}
\alias{lcmmtp}
\title{Causal Mediation with Longitudinal Data Using Modified Treatment Policies}
\usage{
lcmmtp(data, vars, d_prime, d_star, learners, folds)
}
\arguments{
\item{data}{A data frame containing all necessary variables}

\item{vars}{An `lcmmtp_variables` object mapping observed variables to
the assumed variable structure.}

\item{folds}{The number of folds for cross-fitting}
}
\value{
An object of class `lcmmtp`
}
\description{
Causal Mediation with Longitudinal Data Using Modified Treatment Policies
}
