% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmmtp.R
\name{lcmmtp}
\alias{lcmmtp}
\title{Title}
\usage{
lcmmtp(
  data,
  treatment,
  outcome,
  mediator,
  competingRisks,
  baselineConfounders,
  timeVaryConfounders,
  mediatorOutcomeConfounders,
  censoring,
  d_prime,
  d_star,
  id = NULL,
  control = .lcmmtp_control()
)
}
\arguments{
\item{control}{}
}
\description{
Title
}
